+++ src/app/app-routing.module.ts
@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
+import { RouterModule, Routes, UrlSegment } from '@angular/router';
 
 import { PageNotFoundComponent } from './pages/page-not-found/page-not-found.component';
 import { LINK_INFOS } from './global-variables';
@@ -9,27 +9,43 @@ let animationCount = 0;
 const getRouteStateName = () => 'a-' + animationCount++;
 
 let flatLinkInfos = (arr: LinkInfo[]) => {
-  let tmpRoutes: Routes = [];
+  let rtnRoutes: Routes = [];
 
   arr.forEach((linkInfo, index) => {
-    tmpRoutes.push({
+    rtnRoutes.push({
       path: linkInfo.url,
       component: linkInfo.component,
       data: { animation: getRouteStateName() }
     });
 
     if (linkInfo.subLink !== undefined) {
-      let t = flatLinkInfos(linkInfo.subLink);
-      tmpRoutes = tmpRoutes.concat(t);
+      let tmpRoutes = flatLinkInfos(linkInfo.subLink);
+      rtnRoutes = rtnRoutes.concat(tmpRoutes);
     }
   });
 
-  return tmpRoutes;
+  return rtnRoutes;
 }
-let tmpRoutes: Routes = flatLinkInfos(LINK_INFOS);
+let flatRoutes: Routes = flatLinkInfos(LINK_INFOS);
 
 const routes: Routes = [
-  ...tmpRoutes,
+  {
+    matcher: (url) => {
+      //console.log(url);
+      if (url.length > 0 && url[0].path.match(/hide/)) {
+        return {
+          consumed: [url[0]],
+          posParams: {
+            hideLayout: new UrlSegment("hide", {})
+          }
+        };
+      }
+
+      return null;
+    },
+    children: [...flatRoutes]
+  },
+  ...flatRoutes,
   { path: '', redirectTo: '/investment/versus', pathMatch: 'full' },
   { path: '404', component: PageNotFoundComponent, data: { animation: getRouteStateName() } },
   { path: '**', redirectTo: '/404' }